% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgm-auth.R
\name{bgm_auth}
\alias{bgm_auth}
\title{Authentication for Bangumi.tv}
\usage{
bgm_auth(
  new_user = FALSE,
  key = "bgm2725b1fa2fea5582",
  secret = "b84bdc8fcc3c68dd277a6b9650f33867",
  cache = TRUE,
  rdstoken = NA
)
}
\arguments{
\item{new_user}{Set to \code{TRUE} if you need to switch to a new user
account or just flush existing token. Default is \code{FALSE}.}

\item{key}{Your application key (ID). \code{bangumir} already comes with a
key/secret but you are welcome to swap out with our own. Since these keys are
shipped with the package, there is a small chance they could be voided if
someone abuses the key. If you plan to use this in production, or for an
internal tool, the recommended practice is to create a new application on
\href{https://bgm.tv/dev/app}{Bangumi} and use those keys for your purposes.}

\item{secret}{Your application secret. Like \code{key}, \code{bangumir} comes
with a secret but you are welcome to swap out with our own.}

\item{cache}{By default your credentials are locally cached in a file called
\code{.httr-oauth}. Set to \code{FALSE} if you need to authenticate
separately each time.}

\item{rdstoken}{File path to stored RDS token. In server environments where
interactive OAuth is not possible, a token can be created on a desktop client
and used in production. See examples.}
}
\value{
A Token2.0 object, invisibly
}
\description{
This function authenticates you into Bangumi. The documentation for the
\href{https://bangumi.github.io/api/}{Bangumi API}
provides more details including alternate methods if you desire to
reimplement your own.
}
\examples{
\dontrun{

# To either read token from .httr-oauth in the working directory or open a
# web browser to authenticate (and cache a token)
bgm_auth()

# If you want to overwrite an existing local token and switch to a new
# user, set new_user to TRUE.
bgm_auth(new_user = TRUE)

# To store a token for re-use (more flexible than .httr-oauth), save the
# output of bgm_auth and save it to an RDS file
token <- bgm_auth()
saveRDS(token, "/path/to/tokenfile.RDS")

# To use a stored token provide token location
bgm_auth(rdstoken = "/path/to/tokenfile.RDS")
}
}
\references{
\href{https://bangumi.github.io/api/}{API documentation}
}
