% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgm-auth.R
\name{bgm_get_token}
\alias{bgm_get_token}
\title{Retrieve oauth2 token from bangumir-namespaced environment}
\usage{
bgm_get_token()
}
\description{
Retrieves a token if it is previously stored, otherwise prompts user to get
one.
}
\keyword{internal}
