% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgm-latest.R
\name{bgm_latest}
\alias{bgm_latest}
\title{Get summary of user's collection}
\usage{
bgm_latest(
  username,
  type = c("anime", "book", "game", "music", "real"),
  max = 25,
  format = c("list", "table", "default"),
  key = "bgm2725b1fa2fea5582"
)
}
\arguments{
\item{username}{User name a user id.}

\item{type}{Category type, \code{anime}, \code{book}, \code{game},
\code{music} and \code{real} are available.}

\item{max}{Max result number, default is the \code{25} (Maximum value).}

\item{format}{Result will be a list of json data when set to \code{list}
(default), \code{response()} object when set to \code{default}, and list of
\code{data.table} when set to \code{table}.}

\item{key}{Your application key (ID). \code{bangumir} already comes with a
key/secret but you are welcome to swap out with our own. Since these keys are
shipped with the package, there is a small chance they could be voided if
someone abuses the key. If you plan to use this in production, or for an
internal tool, the recommended practice is to create a new application on
\href{https://bgm.tv/dev/app}{Bangumi} and use those keys for your purposes.}
}
\value{
A list of response content, \code{response()} object, or a
\code{data.table} contains the latest collections.
}
\description{
Get summary of user's collection
}
\examples{
result <- bgm_latest(username = "sai", format = "table")

# Result of watching
result$do
}
