% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgm-progress.R
\name{bgm_progress}
\alias{bgm_progress}
\title{Get user's progress}
\usage{
bgm_progress(
  subject_id = NULL,
  format = c("list", "table", "default"),
  token = bgm_get_token()
)
}
\arguments{
\item{subject_id}{Get progress information for a given subject.}

\item{format}{Result will be a list of json data when set to \code{list}
(default), \code{response()} object when set to \code{default}, and a
\code{data.table} when set to \code{table}.}

\item{token}{Bangumi access token which is generated by \code{bgm_auth()}.}
}
\value{
A list of response content, \code{response()} object, or a
\code{data.table} contains the progress.
}
\description{
Get user's progress
}
\examples{
\dontrun{
bgm_progress()
}
}
