% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgm-subject.R
\name{bgm_subject}
\alias{bgm_subject}
\title{Get subject information.}
\usage{
bgm_subject(
  id = NULL,
  responseGroup = c("small", "medium", "large"),
  format = c("list", "default")
)
}
\arguments{
\item{id}{Subject id.}

\item{responseGroup}{Response of details type is set to be \code{small}
(simplified mode), \code{medium} and \code{large} are also available.}

\item{format}{Result will be a list of json data when set to \code{list}
(default), \code{response()} object when set to \code{default}.}
}
\value{
A list of response content or a \code{response()} object according to
the \code{format} option.
}
\description{
Get subject information.
}
\examples{
bgm_subject(10380)
}
