% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgm-subject.R
\name{bgm_subject_ep}
\alias{bgm_subject_ep}
\title{Get subject episode information.}
\usage{
bgm_subject_ep(id = NULL, format = c("list", "table", "default"))
}
\arguments{
\item{id}{Subject id.}

\item{format}{Result will be a list of json data when set to \code{list}
(default), \code{response()} object when set to \code{default}. Set to
\code{table} will convert the episode list to a \code{data.frame} for ease of
use.}
}
\value{
A list of response content or a \code{response()} object according to
the \code{format} option.
}
\description{
Get subject episode information.
}
\examples{
bgm_subject_ep(10380, "table")
}
