% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgm-user-collection.R
\name{bgm_user_collection}
\alias{bgm_user_collection}
\title{Get user's collection}
\usage{
bgm_user_collection(
  username,
  cat = c("watching", "all_watching"),
  ids = NULL,
  responseGroup = c("medium", "small"),
  format = c("list", "table", "default")
)
}
\arguments{
\item{username}{User name a user id.}

\item{cat}{Collection category, \code{watching} for anime and real,
\code{all_watching} for anime, real and book.}

\item{ids}{Numeric vector of subject id.}

\item{responseGroup}{Response of details type is set to be \code{medium},
setting \code{small} for simplified mode.}

\item{format}{Result will be a list of json data when set to \code{list}
(default), \code{response()} object when set to \code{default}, and a
\code{data.table} when set to \code{table}.}
}
\description{
Get user's collection
}
\examples{
bgm_user_collection(username = "sai", format = "table")
}
