% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgm-user-info.R
\name{bgm_user_info}
\alias{bgm_user_info}
\title{Get bangumi user information}
\usage{
bgm_user_info(username, format = "list", force = FALSE)
}
\arguments{
\item{username}{A vector of user name or id.}

\item{format}{Result will be a list of json data when set to \code{l} or
\code{list} (default), a list of \code{response()} object when set to
\code{d} or \code{default}, and a \code{data.frame} when set to \code{t} or
\code{table}.}

\item{force}{By default, you can get information of up to \code{10} users at
a time, but if you want to break the limit, set it to \code{TRUE} but it's
not recommended.}
}
\value{
A list of json data in list, \code{response()} object, or a
\code{data.frame} contains all users' information.
}
\description{
Get bangumi user information
}
\examples{
library(bangumir)
bgm_user_info(username = "sai")
bgm_user_info(username = c("sai", "infinityloop"), format = "t")
}
